Better image stuff

+++ docker/build-version.sh
@@ -6,11 +6,24 @@ export PY_ROOT="/opt/python/${1}/bin"
 
 export PATH=${PY_ROOT}:${PATH}
 
+CLEAN=1
+
+while getopts "n" arg
+do
+    case $arg in
+        n) CLEAN=0 ;;
+    esac
+done
+
 cd /src
-echo "Clearing Build Directory"
-rm -rf build
-echo "Clearing Dist Directory"
-rm -rf dist
+if [ CLEAN = 1 ]
+then
+    echo "Clearing Build Directory"
+    rm -rf build
+    echo "Clearing Dist Directory"
+    rm -rf dist
+fi
+
 echo "Using python $(which python)"
 python -m build --wheel
 echo "Auditing"
